move to dashboard path for admin dashboard
