+++ src/app/dashboard/page.tsx
@@ -1,0 +1,12 @@
+'use client';
+import Layout from "@/components/Layout";
+
+export default function HomeDashboard() {
+  return (
+    <Layout>
+      <div style={{ display: 'flex', flex: 11, justifyContent: 'center', alignItems: 'center' }}>
+        <b>Chart</b>
+      </div>
+    </Layout>
+  );
+}
