+++ src/app/dashboard/products/useDeleteProduct.ts
