+++ src/app/page.tsx
@@ -1,12 +1,6 @@
-'use client';
-import Layout from "@/components/Layout";
 
 export default function Home() {
   return (
-    <Layout>
-      <div style={{ display: 'flex', flex: 11, justifyContent: 'center', alignItems: 'center' }}>
-        <b>Chart</b>
-      </div>
-    </Layout>
+    <div>Home Page</div>
   );
 }
