+++ src/constants/auth.ts
@@ -1,0 +1,2 @@
+
+export const DASHBOARD_LOGIN_PATH = '/dashboard/login'
